requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
connexion==2.14.2
dnspython==2.8.0
email-validator==2.3.0
Flask==2.3.3
flask-cors==6.0.1
Flask-Login==0.6.3
flask-marshmallow==0.15.0
Flask-Principal==0.4.0
Flask-Security-Too==5.3.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.2.2
greenlet==3.2.4
idna==3.11
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==3.21.2
marshmallow-sqlalchemy==0.28.0
packaging==25.0
passlib==1.7.4
pip==25.3
psycopg2-binary==2.9.11
reportlab==4.2.2
setuptools==80.9.0
SQLAlchemy==1.4.46
Werkzeug==3.1.3
WTForms==3.2.1
